% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Application Server Logic}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal Shiny parameters.}
}
\value{
Called for side effects. Returns \code{NULL} invisibly.
}
\description{
Application Server Logic
}
\keyword{internal}
