% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clonotypes}
\alias{clonotypes}
\title{Get specific clonotype data}
\usage{
clonotypes(allData, allele, gene, junction, name, run_diagnosis)
}
\description{
Count unique 'IMGT.gapped.nt.sequences.CDR3.IMGT'
}
\details{
Create convergent evolution list

Calculate clonotype frequency

Get specific clonotype data

Count unique 'IMGT.gapped.nt.sequences.CDR3.IMGT'

Create coergenet evolution list

Calculate clonotype frequency

Get V genes and AA junctions
Get specific clonotype data

Count unique 'IMGT.gapped.nt.sequences.CDR3.IMGT'

Create convergent evolution

Calculate clonotype frequency

Get specific clonotype data

Count unique 'IMGT.gapped.nt.sequences.CDR3.IMGT'

Create convergent evolution list
Get V gene and AA junction

Calculate clonotype frequency

Printing clonotypes file

Printing whole integrated clonotypes file

Printing SHM file

Printing diagnosis file
}
