% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_final_acs_data.R
\name{create_final_acs_data}
\alias{create_final_acs_data}
\title{Prepare acs data for regression model}
\usage{
create_final_acs_data(
  combined_acs_data,
  intersecting_tracts,
  start_month,
  end_month
)
}
\arguments{
\item{combined_acs_data}{output from the \code{combine_acs_data()} function}

\item{intersecting_tracts}{output from the \code{create_intersecting_tract_percentages()} function}

\item{start_month}{the year and month (as a string separated by "-") that the final output will start with. Example: "2021-01"}

\item{end_month}{the year and month (as a string separated by "-") that the final output will end with. Example: "2025-12"}
}
\value{
a data frame with the needed variables to go into the regression model
}
\description{
Prepare acs data for regression model
}
