% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_intersecting_tract_percentages.R
\name{create_intersecting_tract_percentages}
\alias{create_intersecting_tract_percentages}
\title{Find tracts that intersect the route buffer}
\usage{
create_intersecting_tract_percentages(tract_geom, route_geom)
}
\arguments{
\item{tract_geom}{the geometry of the census tracts that
was output from the \code{get_tract_geometry} function.}

\item{route_geom}{the route geometry that was output
from the \code{get_gtfs_routes} function.}
}
\value{
a data frame with geometries of all the intersections created
when crossing the tract boundaries with the route buffer. It
also includes columns for the percentage of the tract it takes up
}
\description{
Find tracts that intersect the route buffer
}
