% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_regression_model.R
\name{create_regression_model}
\alias{create_regression_model}
\title{Create regression model to predict bus ridership}
\usage{
create_regression_model(
  data_xlsx,
  acs_data,
  gas_data,
  variables,
  fare_df = NULL
)
}
\arguments{
\item{data_xlsx}{Agency-provided excel data as an r object of class \code{df}}

\item{acs_data}{Organized ACS data pulled from the census API using a few functions
like \code{pull_acs_data()} and \code{create_final_acs_data()}}

\item{gas_data}{Internal data storing historical gas prices obtained
from \href{https://www.eia.gov/dnav/pet/pet_pri_gnd_dcus_nus_m.htm}{U.S. Energy Information Administration}}

\item{variables}{A vector of variable names that will be tried in the regression model.}

\item{fare_df}{An optional data frame that describes the dates of the fare changes.
Default is NULL.}
}
\value{
Results of the regression model. An R oject of class "fixest"
}
\description{
This is a main function of the TRiP tool. It uses agency-level ridership data and external
census data to create a fixed effects regression model that can help
predict bus ridership. This function uses a backwards stepwise method
to create a model with only coefficients that have a p-value less than 0.1.
}
