% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_overlapping_counties.R
\name{find_overlapping_counties}
\alias{find_overlapping_counties}
\title{Find the counties that overlap the route geometry.}
\usage{
find_overlapping_counties(route_geom)
}
\arguments{
\item{route_geom}{The route geometry that was created with \code{get_gtfs_routes()}}
}
\value{
A data frame with the following columns, "STATEFP", "COUNTYFP", "NAME"
}
\description{
This uses the county shape data found at
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}{census.gov}
}
