% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_acs_year.R
\name{get_max_acs_year}
\alias{get_max_acs_year}
\title{Find most recently available year for ACS 5-year annual estimates}
\usage{
get_max_acs_year()
}
\value{
The most recent year the ACS 5-year estimates
}
\description{
Find most recently available year for ACS 5-year annual estimates
}
\examples{
get_max_acs_year()
}
