% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tract_geometry.R
\name{get_tract_geometry}
\alias{get_tract_geometry}
\title{Get census tract geometry for given counties}
\usage{
get_tract_geometry(state_fips, county_fips, years)
}
\arguments{
\item{state_fips}{A state fips or vector of state fips}

\item{county_fips}{A county fips or vector of county fips}

\item{years}{A year or list of years}
}
\value{
An sf object of the requested tract boundaries that have been validated using \code{st_make_valid}
}
\description{
This function uses the tirgis package to retrieve the census tract boundaries.
}
