% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{homedist}
\alias{homedist}
\title{Calculate maximum distance from 'home' for each trip}
\usage{
homedist(x, home = NULL)
}
\arguments{
\item{x}{trip object}

\item{home}{see details}
}
\value{
numeric vector of distances in km (for longlat), or in the units of the trip's projection
}
\description{
This function returns a distance from a given 'home' coordinate for each individual trip.
Use the \code{home} argument to provide a single, common 2-element (x,y or lon,lat) coordinate. If \code{home}
is \code{NULL} (the default), then each individual trip's first location is used.
}
\seealso{
\code{\link[sp]{spDistsN1}}
}
