% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_data_frame.R
\name{make_model_data_frame}
\alias{make_model_data_frame}
\title{list variables to be log transformed
three variables not log transformed: "workers_16_over", "median_earnings", "emp_pop_ratio"}
\usage{
make_model_data_frame(data_xlsx, acs_data, gas_data, fare_df = NULL)
}
\description{
list variables to be log transformed
three variables not log transformed: "workers_16_over", "median_earnings", "emp_pop_ratio"
}
