% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plat forecasted UPT}
\usage{
plot_forecast(forecast_df, route = "all_routes", scale = "average")
}
\arguments{
\item{forecast_df}{The data frame with forecasted transit trips created from
the forecast_ridership() function}

\item{route}{the name of the route you want to see. It must exist in the forecast_df}

\item{scale}{either "average" or "total" to specify whether you want the y axis to have
average weekday UPT or total weekday UPT for the month.}
}
\description{
Plat forecasted UPT
}
