% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_acs_data.R
\name{pull_acs_data}
\alias{pull_acs_data}
\title{Use to county data frame to get needed acs tables}
\usage{
pull_acs_data(county_sf, years = 2024)
}
\arguments{
\item{county_sf}{the county sf object that was output by \code{find_overlapping_counties()}}

\item{years}{the years that you want the tables for}
}
\value{
a nested list with acs data for each year and each table we want (tables: "B08006","B08201", "DP05","S0802","S2301")
}
\description{
Use to county data frame to get needed acs tables
}
