% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grid.R
\name{rasterize}
\alias{rasterize}
\title{Rasterize trip objects based on line-segment attributes.}
\arguments{
\item{x}{\code{trip} object}

\item{y}{Raster* object}

\item{field}{attribute from which differences will be calculated, defaults to the time-stamp between trip locations}
}
\value{
RasterLayer
}
\description{
Trip rasterize.
}

