% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproj.R
\name{reproj}
\alias{reproj}
\alias{reproj.trip}
\title{Reprojection}
\usage{
\method{reproj}{trip}(x, target, ..., source = NULL)
}
\arguments{
\item{x}{trip object}

\item{target}{target projection}

\item{...}{ignored}

\item{source}{projection of source data, usually ignore this for trips}
}
\value{
a trip reprojected to 'target'
}
\description{
A reproj method for trip objects.
}
