% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}, such as \code{options}.}
}
\value{
A Shiny app object. Called for its side effect of launching the app.
}
\description{
Launches the application in the user's default browser.
}
\examples{
\dontrun{
run_app()
}

}
