\name{spTransform-methods}
\docType{methods}
\title{ Methods for Function spTransform in package "trip" }
\alias{spTransform-methods}
\alias{spTransform}
\alias{spTransform,ANY-method}
\alias{spTransform,trip,CRS-method}
%\alias{spTransform,SpatialPointsDataFrame,CRS-method}

\section{Methods}{\describe{

%\item{"ANY"}{default void method}

\item{"trip", CRSobj = CRS}{returns transformed coordinates of a "trip" object using the projection arguments in "CRSobj", of class CRS}

}}

\description{Transformation in package "trip"  uses the facilities of the rgdal and sp packages.}

\seealso{
 Packages rgdal and sp for full documentation. 
}



\keyword{methods}
\keyword{spatial}
