% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-aesthetics.R
\name{trip2}
\alias{trip2}
\title{specify trip by standard grammar}
\usage{
trip2(.data, ..., crs = NULL)
}
\arguments{
\item{.data}{data frame}

\item{...}{named expressions in the form 'x = longitude'}

\item{crs}{projection in PROJ.4 form}
}
\value{
trip object
}
\description{
Construct 'trip' objects from user-nominated variables, grouping and arrangement. 
Currently hardcoded to require 'x', 'y', 'time' and 'id'.
}
\details{
longitude latitude is assumed if data is in reasonable range of that domain.
}
\examples{
d <- tibble::as_tibble(as.data.frame(walrus818[1:100, ]))
trip2(d, x = X_AED170_70, y = Y_AED170_70, time = DataDT, id = Deployment)
}
