\name{ppp-class}
\docType{class}
\alias{ppp-class}
\alias{owin-class}
\alias{psp-class}
\alias{ltraj-class}

\title{Virtual classes for coercion to and from trip objects}
\description{Virtual S4 class definition for S3 classes in the spatstat
  and adehabitatLT packages to allow S4-style coercion to these classes}
\section{Objects from the Class}{Virtual Classes: No objects may be
  created from these}
\author{Michael D. Sumner}
\keyword{classes}