% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip-package.R
\docType{data}
\name{walrus818}
\alias{walrus818}
\title{Walrus tracking data set.}
\description{
Behavior of Pacific Walruses Tracked from the Alaska Coast of the Chukchi Sea.
}
\details{
Data set is provided as a 'trip' object. This is the abstract for the work:

"We tracked movements and haulout foraging behavior of walruses instrumented with
satellite-linked data loggers from the Alaskan shores of the Chukchi Sea during the
autumn of 2009 (n=13) and 2010 (n=2)."
Jay, C. V. and Fischbach, A.S.
}
\examples{
data(walrus818)
plot(walrus818)
lines(walrus818)

##dontdoanything
## library(mapview)
##mapview(as(walrus818, "SpatialLinesDataFrame"), burst = TRUE)
}
