% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip-package.R
\docType{data}
\name{world_north}
\alias{world_north}
\title{World north polygons}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 185 rows and 11 columns.
}
\usage{
world_north
}
\description{
A spatial polygons object with coastlines of the northern hemisphere.
}
\details{
This data set exists purely to avoid requiring reprojection in the
vignette, the data uses the same projection as \link{walrus818}.
}
\keyword{datasets}
