% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coefficients.R
\name{check_coefficients}
\alias{check_coefficients}
\title{Check model coefficient reasonableness}
\usage{
check_coefficients(model, extra_vars)
}
\arguments{
\item{model}{a model created to forecast bus ridership in the TRiP app.}
}
\value{
A gt table of coefficients that is colored to point out unexpected signs
}
\description{
Check model coefficient reasonableness
}
