% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_acs_data.R
\name{combine_acs_data}
\alias{combine_acs_data}
\title{Organize Pulled ACS Data}
\usage{
combine_acs_data(pulled_acs_list)
}
\arguments{
\item{pulled_acs_list}{The nested list created by \code{pull_acs_data()}}
}
\value{
A combined list of ACS variables with a column for each variable
}
\description{
Organize Pulled ACS Data
}
