% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip.R
\name{forceCompliance}
\alias{forceCompliance}
\title{Function to ensure dates and times are in order with trip ID}
\usage{
forceCompliance(x, tor)
}
\arguments{
\item{x}{\code{\link{data.frame}} or
\code{\link[sp]{SpatialPointsDataFrame-class}}}

\item{tor}{character vector of names of date-times and trip ID columns}
}
\value{
\code{\link{data.frame}} or
\code{\link[sp]{SpatialPointsDataFrame-class}}.
}
\description{
A convenience function, that removes duplicate rows, sorts by the date-times
within ID, and removes duplicates from a data frame or
SpatialPointsDataFrame.
}
\note{
It's really important that data used are of a given quality, but this
function makes the most common trip problems easy to apply.
}
\seealso{
\code{\link{trip}}
}
