% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_ridership.R
\name{forecast_ridership}
\alias{forecast_ridership}
\title{add scenarios to forecast_grid
and get elasticity factors}
\usage{
forecast_ridership(
  coefs,
  data_xlsx,
  acs_data,
  gas_csv,
  scenario_inputs_df,
  start_year = NULL,
  start_month = NULL,
  fare_df = NULL,
  brt_df = NULL
)
}
\description{
add scenarios to forecast_grid
and get elasticity factors
}
