% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtfs_routes.R
\name{get_gtfs_routes}
\alias{get_gtfs_routes}
\title{Get transit route geometry}
\usage{
get_gtfs_routes(gtfs_zip_file)
}
\arguments{
\item{gtfs_zip_file}{The gtfs zip file}
}
\value{
The route geometry from the file
}
\description{
Get transit route geometry
}
