% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{oc.theme}
\alias{oc.colors}
\alias{oc.theme}
\title{SeaWiFS ocean colour colours}
\usage{
oc.theme(x = 50)

oc.colors(n)
}
\arguments{
\item{x}{Number of colours to generate as part of a theme}

\item{n}{Number of colours to generate}
}
\value{
A set of colours or a theme object.
}
\description{
Generate ocean colour colours, using the SeaWiFS scheme
}
\details{
This is a high-contrast palette, log-scaled originally for ocean
chlorophyll.
}
\examples{
\dontrun{
 oc.colors(10)
 library(lattice)
 trellis.par.set(oc.theme())
 utils::example("trip-methods", package="trip",
                ask=FALSE, echo=FALSE)
 tg <- tripGrid(tr)
 spplot(tg)
}
}
\seealso{
Similar functions in sp \code{\link[sp]{sp.theme}},
\code{\link[sp]{bpy.colors}}
}
\keyword{color}

