% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast_facet.R
\name{plot_forecast_facet}
\alias{plot_forecast_facet}
\title{Plot forecasted UPT for all routes in a facet wrap}
\usage{
plot_forecast_facet(forecast_df, scale = "average")
}
\arguments{
\item{forecast_df}{The data frame with forecasted transit trips created from
the forecast_ridership() function}

\item{scale}{either "average" or "total" to specify whether you want the y axis to have
average weekday UPT or total weekday UPT for the month.}
}
\description{
Plot forecasted UPT for all routes in a facet wrap
}
