% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{TimeOrderedRecords-class}
\alias{TimeOrderedRecords-class}
\title{A class for the identifiers of DateTime and ID records in spatial data.}
\description{
The main use of this class and creator function is for
\code{\link[sp]{SpatialPointsDataFrame-class}}s which are used with
TimeOrderedRecords for the class \code{trip}.
}
\note{
Future versions may change significantly, this class is
very basic and could probably be implemented in a better
way. Specifying TOR columns by formula would be a useful addition.
}
\section{Slots}{

\describe{
\item{\code{TOR.columns}:}{2-element vector of class \code{"character"}}
}
}

\examples{
showClass("TimeOrderedRecords")
tor <- new("TimeOrderedRecords", TOR.columns=c("datetime", "ID"))
}
\seealso{
\code{\link{TimeOrderedRecords}}, \code{\link{trip}} for creating trip objects, and \code{\link{trip-class}} for that class
}
