%% $Id$
\name{trip-deprecated}

\alias{trip.split.exact}

% -------------------------------------------------------------------------

\title{Deprecated functions in trip}


\description{
  These functions will be declared defunct in a future release.
}


\usage{
trip.split.exact(x, dates)
}


\arguments{

  \item{x}{see \code{\link{cut.trip}}}

  \item{dates}{see \code{\link{cut.trip}}}
  
}


\details{

  See \code{\link{cut.trip}}.

}


\seealso{

  \code{\link{cut.trip}}

}

% -------------------------------------------------------------------------

\keyword{manip}
