%% $Id$
\name{trip-deprecated}
\alias{trip-deprecated}

\alias{trip.split.exact}
\alias{as.trip.SpatialLinesDataFrame}

% -------------------------------------------------------------------------

\title{Deprecated functions in trip}


\description{
  These functions will be declared defunct in a future release.
}


\usage{
trip.split.exact(x, dates)
as.trip.SpatialLinesDataFrame(from)
}


\arguments{

  \item{x}{see \code{\link{cut.trip}}}

  \item{dates}{see \code{\link{cut.trip}}}

  \item{from}{see \code{\link{as.SpatialLinesDataFrame.trip}}}

}


\seealso{

  \code{\link{cut.trip}}, \code{\link{as.SpatialLinesDataFrame.trip}}

}

% -------------------------------------------------------------------------

\keyword{manip}
