\name{trip-internal}
\alias{trip-internal}
\alias{validtordata}
\alias{trip.default}
\alias{dim.trip}
\alias{[[.trip}
\alias{[[<-.trip}
\alias{$.trip}
\alias{$<-.trip}
%\alias{split.data.frame}
\alias{subset.trip}
\alias{summary.tordata}
\alias{print.summary.tordata}
\alias{print.trip}
\alias{recenter.trip}
\alias{oc.col}
\alias{single.trip.split}

%% temporarily added to pass CHECK, before real hard NAMESPACE work
%% 2012-10-09
\alias{g2ow}
\alias{gcdist.c}
\alias{intpFun}
\alias{penalized}
\alias{tripRbind}


%\alias{spTransform,trip-method}
\title{Internal trip Functions}
\description{
  Internal trip functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written, or for me to work out
  how NAMESPACES really work :).
}
\keyword{internal}
