%% $Id$
\name{trip-internal}
\alias{trip-internal}

\alias{.validTORdata}
\alias{.oc.col}
\alias{.single.trip.split}
\alias{.g2ow}
\alias{.gcdist.c}
\alias{.intpFun}
\alias{.tripRbind}
\alias{.distances}
\alias{.abdali}


\title{Internal trip Functions}


\description{
  Internal trip functions
}


\details{
  
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written). 
}


\keyword{internal}
