%% $Id: trip-internal.Rd 115 2013-04-25 17:26:50Z sluque $
\name{trip-internal}
\alias{trip-internal}

\alias{.validTORdata}
\alias{.oc.col}
\alias{.single.trip.split}
\alias{.g2ow}
\alias{.gcdist.c}
\alias{.intpFun}
\alias{.tripRbind}
%\alias{spTransform,trip-method}


\title{Internal trip Functions}


\description{
  Internal trip functions
}


\details{
  
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written, or for me to work out
  how NAMESPACES really work :).
  
}


\keyword{internal}
