% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R, R/algorithms.R, R/angle.R,
%   R/colours.R, R/distance.R, R/grid.R
\docType{data}
\name{trip-internal}
\alias{trip-internal}
\alias{.validTORdata}
\alias{.oc.col}
\alias{.single.trip.split}
\alias{.g2ow}
\alias{.gcdist.c}
\alias{.intpFun}
\alias{.tripRbind}
\alias{.distances}
\alias{.abdali}
\title{Internal trip Functions}
\format{
An object of class \code{character} of length 256.
}
\usage{
.validTORdata(object)

.intpFun(x)

.abdali(p1, p2)

.oc.col

.distances(x)

.g2ow(x)
}
\description{
Internal trip functions
}
\details{
These are not to be called by the user (or in some cases are just waiting
for proper documentation to be written).
}
\keyword{datasets}
\keyword{internal}
