\name{trip-package}
\alias{trip-package}
\docType{package}
\title{
Spatial analysis of animal track data
}
\description{
Functions for accessing and manipulating spatial data for animal tracking.
   Filter for speed and create time spent plots from animal track data.
   Classes \code{TimeOrderedRecords} and \code{trip} provide an extension to the
   \code{\link[sp]{Spatial}} classes.
}
\details{
\tabular{ll}{
Package: \tab trip\cr
Type: \tab Package\cr
Version: \tab 0.0-4\cr
Date: \tab 2006-08-10\cr
License: \tab GPL2\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Michael D. Sumner

Maintainer: Michael D. Sumner <mdsumner@utas.edu.au>

}
\references{
%~~ Literature or other references for background information ~~
}
\keyword{ package }

\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
