% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRiP_package.R
\docType{package}
\name{TRiP-package}
\alias{TRiP}
\alias{TRiP-package}
\title{TRiP: Transit Ridership Prediction Tool}
\description{
A Shiny application that forecasts bus ridership for small- to medium-sized
transit agencies. TRiP combines historical route-level ridership with service
levels, demographic and economic indicators, and telecommuting rates to fit a
fixed-effects regression model and project ridership forward.
}
\section{Main workflow}{

The app proceeds through the following stages, each backed by functions in
this package:
\enumerate{
\item Load and screen agency ridership data
\item Attach GTFS-derived service characteristics
\item Join ACS demographics and gas price series
\item Fit the regression model
\item Forecast and visualize future ridership
}
}

\section{Getting started}{

Launch the application with \code{\link{run_app}()}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KamrynMansfield/TRiP}
  \item \url{https://kamrynmansfield.github.io/TRiP/}
  \item Report bugs at \url{https://github.com/KamrynMansfield/TRiP/issues}
}

}
\author{
\strong{Maintainer}: Candace Brakewood \email{cvoulgaris@gsd.harvard.edu} (\href{https://orcid.org/0000-0003-2769-7808}{ORCID}) [contributor]

Authors:
\itemize{
  \item Candace Brakewood \email{cvoulgaris@gsd.harvard.edu} (\href{https://orcid.org/0000-0003-2769-7808}{ORCID}) [contributor]
  \item Kamryn Mansfield \email{kmansfi4@vols.utk.edu} (\href{https://orcid.org/0009-0003-5466-0821}{ORCID}) [contributor]
  \item Mtthew Davis [contributor]
}

}
\keyword{internal}
