% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trip-deprecated.R
\name{trip.split.exact}
\alias{as.SpatialLinesDataFrame.trip}
\alias{as.ltraj.trip}
\alias{as.trip.SpatialLinesDataFrame}
\alias{trip-deprecated}
\alias{trip.split.exact}
\alias{tripTransform}
\title{Deprecated functions in trip}
\usage{
as.SpatialLinesDataFrame.trip(from)

trip.split.exact(x, dates)

as.ltraj.trip(xy)

as.trip.SpatialLinesDataFrame(from)

tripTransform(x, crs, ...)
}
\arguments{
\item{from}{trip object}

\item{x}{see \code{\link{cut.trip}}}

\item{dates}{see \code{\link{cut.trip}}}

\item{xy}{\code{trip} object}

\item{crs}{CRS object, or PROJ.4 string accepted by \code{\link[sp]{CRS}}}

\item{\dots}{Further arguments to \code{\link[rgdal]{spTransform}}}
}
\description{
These functions will be declared defunct in a future release.
}
\seealso{
\code{\link{cut.trip}}, \code{\link{as.Other}}
}

