% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip-deprecated.R
\name{trip.split.exact}
\alias{trip.split.exact}
\alias{trip-deprecated}
\alias{as.trip.SpatialLinesDataFrame}
\alias{tripTransform}
\alias{as.ltraj.trip}
\alias{as.SpatialLinesDataFrame.trip}
\title{Deprecated functions in trip}
\usage{
as.SpatialLinesDataFrame.trip(from)

trip.split.exact(x, dates)

as.ltraj.trip(xy)

as.trip.SpatialLinesDataFrame(from)
}
\arguments{
\item{from}{trip object}

\item{x}{see \code{\link{cut.trip}}}

\item{dates}{see \code{\link{cut.trip}}}

\item{xy}{\code{trip} object}
}
\description{
These functions will be declared defunct in a future release.
}
\seealso{
\code{\link{cut.trip}}, \code{\link{as.Other}}
}
